requirements.txt
twine~=4.0.0 # uploader for PyPI
rich~=12.4.1 # rich is also a dependency of twine and pip so it should already be installed.
PyGithub~=1.55 # for creating tags and releases
urllib3~=1.26.9 # for getting the latest version of the package